Fix data types in data collator

+++ language_modelling/data_collator.py
@@ -768,7 +768,7 @@ class DataCollatorForLanguageModeling(DataCollatorMixin):
 
         # 80% of the time, we replace masked input tokens with tokenizer.mask_token ([MASK])
         indices_replaced = torch.bernoulli(torch.full(labels.shape, 0.9)).bool() & masked_indices
-        inputs[indices_replaced] = self.tokenizer.convert_tokens_to_ids(self.tokenizer.mask_token).to(torch.long)
+        inputs[indices_replaced] = int(self.tokenizer.convert_tokens_to_ids(self.tokenizer.mask_token))
 
         # 10% of the time, we replace masked input tokens with random word
         indices_random = torch.bernoulli(torch.full(labels.shape, 0.5)).bool() & masked_indices & ~indices_replaced
@@ -1018,7 +1018,7 @@ class DataCollatorForWholeWordMask(DataCollatorForLanguageModeling):
 
         # 80% of the time, we replace masked input tokens with tokenizer.mask_token ([MASK])
         indices_replaced = torch.bernoulli(torch.full(labels.shape, 0.9)).bool() & masked_indices
-        inputs[indices_replaced] = self.tokenizer.convert_tokens_to_ids(self.tokenizer.mask_token).to(torch.long)
+        inputs[indices_replaced] = int(self.tokenizer.convert_tokens_to_ids(self.tokenizer.mask_token))
 
         # 10% of the time, we replace masked input tokens with random word
         indices_random = torch.bernoulli(torch.full(labels.shape, 0.5)).bool() & masked_indices & ~indices_replaced
